Macaca/requirements.txt
wd==0.1.5
pytest==2.9.0
pytest-allure-adaptor==1.7.6